--- netmanager-app/components/Charts/PMchart.tsx
@@ -1,73 +1,0 @@
-"use client"
-
-import { Pie, PieChart } from "recharts"
-
-import {
-  Card,
-  CardContent,
-  CardHeader,
-} from "@/components/ui/card"
-import {
-  ChartConfig,
-  ChartContainer,
-  ChartLegend,
-  ChartLegendContent,
-} from "@/components/ui/chart"
-const chartData = [
-  { browser: "chrome", visitors: 275, fill: "var(--color-chrome)" },
-  { browser: "safari", visitors: 200, fill: "var(--color-safari)" },
-  { browser: "firefox", visitors: 187, fill: "var(--color-firefox)" },
-  { browser: "edge", visitors: 173, fill: "var(--color-edge)" },
-  { browser: "other", visitors: 90, fill: "var(--color-other)" },
-]
-
-const chartConfig = {
-  visitors: {
-    label: "Visitors",
-  },
-  chrome: {
-    label: "Good",
-    color: "hsl(var(--chart-1))",
-  },
-  safari: {
-    label: "Moderate",
-    color: "hsl(var(--chart-2))",
-  },
-  firefox: {
-    label: "UHFSG",
-    color: "hsl(var(--chart-3))",
-  },
-  edge: {
-    label: "Unhealthy",
-    color: "hsl(var(--chart-4))",
-  },
-  other: {
-    label: "Very Unhealthy",
-    color: "hsl(var(--chart-5))",
-  },
-} satisfies ChartConfig
-
-export function PMchart() {
-  return (
-    <Card className="col-span-12 rounded-sm border border-stroke bg-white px-5 pb-5 pt-7.5 shadow-default dark:border-strokedark dark:bg-boxdark sm:px-7.5 xl:col-span-4">
-      <CardHeader className="items-center pb-0">
-        <h4>AQ Exceedances</h4>
-      </CardHeader>
-      <CardContent className="flex-1 pb-0">
-        <ChartContainer
-          config={chartConfig}
-          className="mx-auto aspect-square max-h-[300px]"
-        >
-          <PieChart>
-            <Pie data={chartData} dataKey="visitors" />
-            <ChartLegend
-              content={<ChartLegendContent nameKey="browser" />}
-              className="-translate-y-2 flex-wrap gap-2 [&>*]:basis-1/4 [&>*]:justify-center"
-            />
-          </PieChart>
-        </ChartContainer>
-      </CardContent>
-    </Card>
-  )
-}
-
